Removed old tree filter component.
