+++ examples/src/TreeFilter.tsx
@@ -3,7 +3,7 @@ import 'ts-helpers';
 
 import * as React from 'react';
 import * as ReactDOM from 'react-dom';
-import { TreeFilter, IFilterSelection, FilterSelectionEnum, TreeFilterNew, TreeFilterCallout } from '../../src/components/TreeFilter';
+import { TreeFilter, IFilterSelection, FilterSelectionEnum, Tree } from '../../src/components/TreeFilter';
 import { createFlatList, createRandomizedData, getSelectedIds } from '../MockData/treeFilterElements';
 
 interface DemoState {
@@ -32,48 +32,27 @@ export class Index extends React.Component<any, DemoState> {
 
         return (
             <div style={{ paddingLeft: 300 }}>
-                <TreeFilterCallout
-                    title="Tree Filter (max size)"
-                    filterId={'f1'}
-                    items={flatList}
-                    onValuesSelected={this.onValuesSelected}
-                    // tslint:disable-next-line:no-string-literal
-                    filterSelection={{
-                        type: FilterSelectionEnum.All,
-                        selectedIDs: []
-                    }}
-                    defaultSelection={FilterSelectionEnum.All}
-                    maxWidth={700}
-                    maxHeight={500}
-                    clearSearchOnClose={false}
-                />
                 <TreeFilter
                     title="Tree Filter (max size)"
-                    filterId={'f10'}
+                    filterId={'f1'}
                     items={treeData}
                     onValuesSelected={this.onValuesSelected}
                     // tslint:disable-next-line:no-string-literal
-                    filterSelection={this.state.filterStates['f10']}
+                    filterSelection={this.state.filterStates['f1']}
                     defaultSelection={FilterSelectionEnum.All}
                     maxWidth={700}
                     maxHeight={500}
-                    clearSearchOnClose={false}
                 />
-                <div className="container" style={{ width: '500px', height: '500px' }}>
-                    <TreeFilterNew
-                        selectionText={(text) => console.log('new selection text : ', text)}
-                        filterId={'f8'}
-                        hasSearch={false}
-                        rowHeight={25}
-                        items={treeData}
-                        onValuesSelected={this.onValuesSelected}
-                        // tslint:disable-next-line:no-string-literal
-                        filterSelection={this.state.filterStates['f8']}
-                        defaultSelection={FilterSelectionEnum.All}
-                    />
-                </div>
-
-                {/* <TreeFilter
+                <TreeFilter
+                    title="Tree Filter - depth 4"
+                    filterId={'f2'}
+                    items={deeperTreeData}
+                    onValuesSelected={this.onValuesSelected}
+                    // tslint:disable-next-line:no-string-literal
+                    filterSelection={this.state.filterStates['f2']}
+                    defaultSelection={FilterSelectionEnum.All}
+                />
+                <TreeFilter
                     title="Single Select"
                     filterId={'f3'}
                     items={treeData}
@@ -93,7 +72,26 @@ export class Index extends React.Component<any, DemoState> {
                         bottomLeft: true,
                         topLeft: false
                     }}
-                /> */}
+                />
+                <TreeFilter
+                    title="Flat list"
+                    filterId={'f4'}
+                    items={flatList}
+                    onValuesSelected={this.onValuesSelected}
+                    itemsAreFlatList={true}
+                    // tslint:disable-next-line:no-string-literal
+                    filterSelection={this.state.filterStates['f4']}
+                />
+                <TreeFilter
+                    title="Flat list - few elements"
+                    filterId={'f5'}
+                    items={shortFlatList}
+                    onValuesSelected={this.onValuesSelected}
+                    itemsAreFlatList={true}
+                    defaultSelection={FilterSelectionEnum.All}
+                    // tslint:disable-next-line:no-string-literal
+                    filterSelection={this.state.filterStates['f5']}
+                />
             </div>
         );
     }
