+++ src/components/TreeFilter/Tree.Props.ts
@@ -1,0 +1,45 @@
+import { FilterSelectionEnum, TreeItem, IFilterSelection } from './TreeFilter.Props';
+import { ItemOperator } from './TreeItemOperators';
+
+const nullFunc = () => { };
+
+export interface ITreeProps {
+    title?: string;
+    hasSearch?: boolean;
+    filterId: string;
+    items: Array<TreeItem>;
+    filterSelection?: IFilterSelection;
+    isSingleSelect?: boolean;
+    isGroupSelectableOnSingleSelect?: boolean;
+    itemsAreFlatList?: boolean;
+    onValuesSelected?: (filterId: string, filterSelection: IFilterSelection) => void;
+    defaultSelection?: FilterSelectionEnum;
+    rowHeight?: number;
+    selectionText?: (selectionText: string) => void;
+    onItemsSearch?: (query: string) => void;
+    searchQuery?: string;
+    allItemIdsGetter?: (items: Array<TreeItem>) => ReadonlyArray<string>;
+    lookupTableGetter?: (items: Array<TreeItem>) => any;
+}
+
+export const defaultTreeProps: Partial<ITreeProps> = {
+    filterId: 'treeFilter',
+    hasSearch: true,
+    isSingleSelect: false,
+    itemsAreFlatList: false,
+    isGroupSelectableOnSingleSelect: false,
+    onValuesSelected: () => { },
+    filterSelection: { type: FilterSelectionEnum.None, selectedIDs: [] },
+    defaultSelection: FilterSelectionEnum.None,
+    rowHeight: 20,
+    onItemsSearch: nullFunc,
+    searchQuery: '',
+    allItemIdsGetter: (items: Array<TreeItem>) => ItemOperator.getAllItemIds(items),
+    lookupTableGetter: (items: Array<TreeItem>) => ItemOperator.getLookupTableAndParentLookup(items)
+};
+
+export interface ITreeState {
+    filteredItems: Array<TreeItem>;
+    searchText: string;
+    partiallyCheckedItemIds: Array<string>;
+}
