+++ src/components/TreeFilter/index.ts
@@ -1,4 +1,3 @@
 export * from './TreeFilter.Props';
 export * from './TreeFilter';
-export * from './TreeFilterNew';
-export * from './TreeFilterCallout';
+export * from './Tree';
